+++ assets/css/style.css
@@ -1,0 +1,24 @@
+/* Base - Variables: tailwind-config.js */
+
+/* 
+  Imports go to the top
+  https://tailwindcss.com/docs/using-with-preprocessors/#build-time-imports 
+*/
+
+/* purgecss start ignore */
+@import "tailwindcss/preflight";
+
+@import "tailwindcss/components";
+
+/* purgecss end ignore */
+
+@import "tailwindcss/utilities";
+
+/* Modules */
+@import "./tailwindcss/modules/_hovers";
+@import "./tailwindcss/modules/_utilities";
+@import "./tailwindcss/modules/_nested";
+
+
+/* Site Specific */
+@import "./site";
