+++ go.mod
@@ -1,10 +1,11 @@
 module github.com/imintel/eick.com
-
+replace github.com/theNewDynamic/hugo-module-tnd-styles => /Users/regisphilibert/Boulot/tnd/hugo-module-tnd-styles
 go 1.14
 
 require (
 	github.com/theNewDynamic/hugo-module-tnd-imgix v0.2.1 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-netlifycms v0.1.5 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-seo v0.1.1 // indirect
+		github.com/theNewDynamic/hugo-module-tnd-styles v0.1.1 // indirect
 	github.com/theNewDynamic/hugo-module-tnd-tags v0.0.0-20200626200509-93dcb309850a // indirect
 )
