+++ layouts/partials/_head/meta-and-links.html
@@ -5,7 +5,7 @@
 <meta name="viewport" content="width=device-width,minimum-scale=1">
 {{ hugo.Generator }}
 
-{{ partial "_head/stylesheets.html" . }}
+{{ partial "tnd-styles/tags.html" . }}
 
 {{ partial "_head/manifest.html" . }}
