+++ layouts/partials/tnd-styles/private/GetStyleAssets.html
@@ -1,0 +1,24 @@
+{{ $style_assets := slice }}
+{{ $return := slice }}
+
+{{ $to_bundle := slice }}
+{{ with resources.GetMatch "css/style.*" }}
+  {{ $style := false }}
+
+  {{ $style = . | resources.PostCSS }}
+  {{ $style_assets = $style_assets | append (dict "style" $style "type" "link") }}
+{{ end }}
+
+{{ with resources.Get "fonts/index.css" }}
+  {{ $style_assets = $style_assets | append (dict "style" . "type" "critical") }}
+{{ end }}
+
+{{ range $style_assets }}
+  {{ $style := . }}
+  {{ if and (not site.IsServer) (ne .type "critical") }}
+    {{ $style = dict "type" .type "style" (.style | minify | fingerprint | resources.PostProcess) }}
+  {{ end }}
+  {{ $return = $return | append $style }}
+{{ end }}
+
+{{ return $return }}
