+++ layouts/partials/tnd-styles/private/PrivatePartialExample.html
@@ -1,0 +1,13 @@
+{{/*
+  PrivvatePartialExample
+
+  @author 
+
+  @context Type (.name)
+
+  @access private
+
+  @example - Go Template
+    {{ partial "path" context }}
+    {{ partialCached "path" context "variant" }}
+*/}}
