+++ package.json
@@ -7,7 +7,8 @@
   "author": "budparr <budparr@gmail.com>",
   "license": "MIT",
   "hugo": {
-    "version": "0.74.3"
+    "version": "0.74.3",
+    "extended": "true"
   },
   "scripts": {
     "cms": "npx netlify-cms-proxy-server",
@@ -26,20 +27,22 @@
     "build:hugo": "env NODE_ENV=production bin/hugo* --gc --minify"
   },
   "devDependencies": {
-    "@fullhuman/postcss-purgecss": "^1.1.0",
-    "autoprefixer": "^9.5.0",
     "intersection-observer": "^0.5.1",
     "lazysizes": "^4.1.7",
     "npm-run-all": "^4.1.5",
     "parcel-bundler": "^1.12.3",
-    "postcss": "^7.0.14",
     "quicklink": "^1.0.0",
-    "tailwindcss": "^0.7.4",
-    "tailwindcss-break": "^0.1.0",
-    "tailwindcss-flexbox-order": "^1.0.0",
-    "tailwindcss-multi-column": "^0.2.0",
     "typeface-play": "^0.0.54",
     "amplitudejs": "^5.0.3"
   },
-  "dependencies": {}
+  "dependencies": {
+    "@fullhuman/postcss-purgecss": "^1.1.0",
+    "autoprefixer": "^9.5.0",
+    "postcss": "^7.0.14",
+    "postcss-import": "^12.0.1",
+    "tailwindcss": "^0.7.4",
+    "tailwindcss-break": "^0.1.0",
+    "tailwindcss-flexbox-order": "^1.0.0",
+    "tailwindcss-multi-column": "^0.2.0"
+  }
 }
